+++ src/main/java/com/liemily/data/Item.java
@@ -2,13 +2,13 @@ package com.liemily.data;
 
 public class Item {
     private final String id;
-    private final int[] propArray;
+    private final double[] propArray;
 
     public Item(final String name) {
-        this(name, new int[0]);
+        this(name, new double[0]);
     }
 
-    public Item(final String name, final int[] propArray) {
+    public Item(final String name, final double[] propArray) {
         this.id = name;
         this.propArray = propArray;
     }
@@ -17,7 +17,7 @@ public class Item {
         return id;
     }
 
-    public int[] getPropArray() {
+    public double[] getPropArray() {
         return propArray;
     }
 }
